test(tests): replace per-file model macros with shared scenarios
field_version.rs and dynamodb_update_batch_filter.rs each defined
inline macro_rules! to share model definitions across tests. Replace
them with scenario! modules under scenarios/, matching the convention
used elsewhere in the suite.

Consolidate the tagged-item shape used by both files into a single
tagged_item scenario carrying tag, status, name, and #[version] —
the version column is inert for the non-version tests, so they share
one model definition without changing semantics.
