+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -2,7 +2,10 @@ pub mod has_many_belongs_to;
 pub mod has_many_multi_relation;
 pub mod has_many_same_target;
 pub mod has_one_optional_belongs_to;
+pub mod tagged_item;
 pub mod two_models;
 pub mod user_unique_email;
 pub mod user_with_age;
+pub mod versioned_item;
+pub mod versioned_user_unique_email;
 pub mod widget_mixed_types;
