+++ crates/toasty-driver-integration-suite/src/scenarios/versioned_user_unique_email.rs
@@ -1,0 +1,20 @@
+use crate::prelude::*;
+
+scenario! {
+    #[derive(Debug, toasty::Model)]
+    struct User {
+        #[key]
+        #[auto]
+        id: uuid::Uuid,
+
+        #[unique]
+        email: String,
+
+        #[version]
+        version: u64,
+    }
+
+    async fn setup(test: &mut Test) -> toasty::Db {
+        test.setup_db(models!(User)).await
+    }
+}
